Update bower.json
`bower install bootstrap` is a very convenient way to install bootstrap into a project. It includes the dist folder for js and css files and also includes the build tools. However by excluding ".*" you are not allowing developers to download the configure files such as less/.csslintrc and grunt/.jshintc. These are crucial elements to building with `grunt` or in general. Its likely best to either only serve the dist folder with a bower install or include all the tools necessary to build bootstrap similar to a git clone.